VISTA-553 Correction to sqft and availableFrom presentation in XML

+++ vista-crm/vista-crm-server/src/main/java/com/propertyvista/crm/server/openapi/model/BuildingRS.java
@@ -44,9 +44,9 @@ public class BuildingRS {
 
     public Double rentTo;
 
-    public Double sqftFrom;
+    public Integer sqftFrom;
 
-    public Double sqftTo;
+    public Integer sqftTo;
 
     public Integer unitCount;
